requirements.txt
beautifulsoup4==4.8.0
img2pdf==0.3.3
requests==2.22.0
Pillow==6.1.0